Refactor: move the COS integration to separate module
